[roll] Roll fuchsia [netstack3] Make device BindingId nonzero
Make BindingId an alias for NonZeroU64 (instead of u64) and use it
instead of the ad-hoc u64 throughout test and non-test code in the
bindings.

Note that this changed the logic of the Devices struct to hold the next
ID to allocate instead of the last ID allocated, though the first ID to
be allocated is still 1.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831517
Original-Revision: decd622971e7fd016efee980abfa2bc462ea42ce
GitOrigin-RevId: 109e1ef3ab6293fe0ab3f72ab0be39c22a853fd4
Change-Id: I0bd49c98fb3c106d4b47ef487ed28240585fc574