[roll] Roll fuchsia [netstack] Deflake Fast UDP unit test
TestDatagramSocketWithBlockingEndpoint/closeAfterUnblocked verifies that
when a socket is closed, it will finish writing any available packets
into the gVisor endpoint (as long as it is not blocked) before cleaning
up. The test, as written, is racy because it does the following:

1. Write packets into the zircon socket
2. Unblock the endpoint
3. Close the socket
4. Expect all packets to be enqueued

And it is possible for the write loop to observe (2) and (3)
simultaneously, in which case there is no guarantee it will finish
enqueueing packets. To deflake this, wait for the write loop to resume
enqueueing packets before initiating socket closure.

Original-Fixed: 123927
Multiply: netstack-gotests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831917
Original-Revision: 4856dff606390d4505ee9b17e655673f3870d4a6
GitOrigin-RevId: fa1c71039bfc8fb6562a3bd21d7ff682958c3247
Change-Id: I1e5b6af889b06f89d2f58c142ad07b150e72df47