[roll] Roll third_party/pigweed pw_presubmit: Fix OwnersError constructor
The previous OwnersError.__init__ stored the message argument in a
custom attribute rather than passing it to Exception.__init__. This
caused str() of the exception to be empty, (or more precisely, repr() to
show just the exception type name), leading to a confusing backtrace:

    20230405 21:07:40 ERR Presubmit check owners_lint_checks failed!
    Traceback (most recent call last):
      ...
      File ".../pw_presubmit/owners_checks.py", line 131, in __init__
	raise OwnersDependencyError(error_msg)
    pw_presubmit.owners_checks.OwnersDependencyError
                                                      ^^^???^^^

The use of Exception.message is obsolete:
https://stackoverflow.com/q/1272138

We can instead omit OwnersError.__init__ and use the base class
Exception.__init__ which accepts a message as the first argument.
This fixes the backtrace:

    20230406 15:11:15 ERR Presubmit check owners_lint_checks failed!
    Traceback (most recent call last):
      ...
      File ".../pw_presubmit/owners_checks.py", line 126, in __init__
	raise OwnersDependencyError(
    pw_presubmit.owners_checks.OwnersDependencyError: Tried to import
      .../OWNERS but it does not exist

This also required replacing `err.message` with `str(err)` -- or
just `err` if it is being formatted with `%s`.

Also fixed a typo: "exists" -> "exist".

Tested: Ran `pw presubmit --step owners_lint_checks` on a repo with a
  missing OWNERS file, and confirmed that the error message which was
  previously missing is now visible.
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/137550
Original-Revision: 56b11ffa2f279e6bfa26af586ed75c4b6f501872
GitOrigin-RevId: 59f9e01d214c285bd93c36a2754464dc06c6a8a9
Change-Id: I6cf270ccb77d3e00d010767409fda834211a60f5