[roll] Roll fuchsia [ffx][config] Make Cache not depend on Environment
This makes the Cache type take the value it stores as a generic so that
the code in that file doesn't need to depend on environmentcontext or
environment at all. This removes a circular dependency within the config
crate that makes it hard to break up into smaller pieces.

Multiply: ffx_config_lib
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828878
Original-Revision: 7f819ed144b681f161ba72bb1a6aa8a9dcc65ead
GitOrigin-RevId: b45b63d4884b373c29e33c61bee18524c2629ca4
Change-Id: I705f1281c2fc7311976c4fd3f62d96e1e47f562c