[roll] Roll fuchsia [bti-test] Wait for file removal
This test causes the driver host to crash and then waits for it
to be re-created.

Specifically wait for the file to be removed from devfs.
This fixes a flake where the driver host has crashed, closing the
client channels, but DriverManager hasn't removed the devfs entry
yet and the test connects to the old entry.

Original-Fixed: 124764
Multiply: platform-bus-bti-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831364
Original-Revision: c27b211f7d78ba024fbc9f54dfdc837352447ce9
GitOrigin-RevId: a9471f62f4d474c4d2b66244c92f0c2a754be0e4
Change-Id: I2a440124dfdf2c215a67c0f77eaf457349c28449