[roll] Roll fuchsia [flatland] Remove extra discard config display calls
Discard in CheckConfig is a costly call. It only discards pending
changes, so it is only necessary if we attempted to set layers on
display. This CL removes the unnecessary calls we make at the start and
when we fallback to gpu compositing due to config.

Original-Bug: 123838, 84271

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831284
Original-Revision: ad29fb067265c9c00542b8b01d6f8e329f0ae6c1
GitOrigin-RevId: 95908e58d195cef0424721d5dea54963734b83ab
Change-Id: Idce5c7d905ebdb797d3d2c14df2d01c1f2f69fda