[roll] Roll fuchsia [driver_host] Release api lock when making sync calls
Holding the API lock when making blocking calls can result in a deadlock
as the driver manager may end up blocked on call back into the driver
host. Releasing the for the duration of the sync call should be safe as
long as the device isn't destroyed, which it shouldn't be as we are
hanging onto a reference.

Original-Bug: b/276771439
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832248
Original-Revision: ece503f2ae95c6d127ee9f686516accf07d7caa3
GitOrigin-RevId: f30e45dd170b01b19ef747aa7899d7d4752ebd2a
Change-Id: Ie0e9b96b078ae0b4f9f597ff9dabecfff1f1166b