[roll] Roll fuchsia [dfv2][compat] Return error synchronously on duplicate sibiling.
In DFv1 this error is synchronously returned, but NodeAdd is called
async in the implementation of `device_add`. Adding a redudndant check
in compat allows us to match behavior.

Additionally a NON_BINDABLE flag was added to the test driver to allow
adding children.

Original-Bug: 124914
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831424
Original-Revision: 10e10995dcc1556a9384af116761984aac8f594c
GitOrigin-RevId: 34da7c862be89bf42935ee31fa1688e0b7eb9cb4
Change-Id: Ic8294733a8fda332aba042f734ceda5a0813be94