[roll] Roll fuchsia [inspect] Fix bug in the `Heap/extend` benchmark
We wereusing `iter_with_large_setup` which internally calls
`iter_batched`. This function moves the input to the function
being measured which leads to measuring the Drop impl of the input.

The correct thing to do here is to use `iter_batched_ref` which
passes a reference to the function executing the code that will be
measured. That way the input is dropped after the function has been
measured and not during.

Original-Bug: 123517

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832276
Original-Revision: 7ca725ed0ce3ef12b5c9fc66836e28e1f6b21948
GitOrigin-RevId: b78d3f27963f85adf08836f7ce33565c368f173e
Change-Id: Id7b0edda272e45221d2cec636443448a17d71aa2