[roll] Roll fuchsia [fx] Add get-flash-source command.
This command is used to retrieve the path to either
a flash manifest, or to a bazel-generated product bundle
when `use_bazel_images_only` is set in the current build
configuration (though this is an implementation detail
that may change in the future).

The output string uses a prefix of either
`flash-manifest:` or `product-bundle:` to distinguish
between these two cases.

This will later be used by `fx flash` and `fx remote-flash`

Original-Bug: 124241

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832500
Original-Revision: 1c78d8934de2800472ef95330c66f8f6d2191c40
GitOrigin-RevId: 9e611d119ac797e5607aab2b0dc691553bac1061
Change-Id: Ifd51a250331bdbd7ae46a7a6750c6e842e1cc1ca