[roll] Roll fuchsia [fuchsia.io] Remove RightsRequest
This was an over-engineered solution to needing POSIX semantics on Open
calls that are not known a priori to be reaching a directory.

The same information is now in the directory variant of NodeProtocols,
which is now a table.

Original-Bug: 77623
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832777
Original-Revision: 4a86ef6e7564b91e1c874760da9c5959fa34e0c7
GitOrigin-RevId: e5acf49eb58eaf463387990c17eb14311faaac75
Change-Id: I80aee85559ba22721b301c2a27f2b8f25f590ff2