[roll] Roll fuchsia [inspect] Exclusive ownership for Container
This refactor makes Container non-Clonable and instead requires a
exclusive reference to it to be able to mutate it (or shared reference
for reading), aligned with Rust ownership semantics.

This is a pre-req for the follow-up which brings a more efficient
Container.

Run-All-Tests: true
Original-Bug: 123517

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821007
Original-Revision: df023e0fc0d8e34de6e817d7a41169b3d08bd1f7
GitOrigin-RevId: 14d5c8e7c528666319e064ee73d4f7229ade8079
Change-Id: I2cb1793fa376a0612a57f88778999e225d275f7e