[roll] Roll fuchsia [driver_runner] Fix overlapping bind calls
Currently, TryBindAllOrphans() and Bind() calls can overlap
with a TryBindAllOrphans() call that is in process.

When these calls overlap, it can cause some orphans to be
excluded from binding. This is because all the calls interact
with |orphaned_nodes_|.

To prevent overlapping, this CL adds a barrier between these
calls so that any Bind() and TryBindAllOrphans() calls are
queued while a TryBindAllOrphans() call is in process.

Once it's completed, it executes the queued calls. This lets
us retry the bind attempts with a complete set of orphaned
nodes.

This CL also fixes a bug in which binding did not report
a successful bind if an orphan binds with a DFv1 style
composite.

Original-Bug: 124988
Tests: device-enumeration-test, driver-runner-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832211
Original-Revision: c399449323d724cffd7ed402fa132c4c8d588a4c
GitOrigin-RevId: 538df970ceabb9ba8c606f35802b4c355c523c5c
Change-Id: Ifb53e48570d64e67b142b955ee8ac1cd81b70024