[roll] Roll fuchsia [build] Add user.basic environment, use it for vDSO
This adds the generalized user.basic environment for building
things to use only the basic machine ABI. This is now used for
the vDSO, replacing the old special user.vdso environment. It
will later be used for things that need to avoid using the
Fuchsia Compiler ABI.

This restores the long-lost feature of doing variant selection
for the vDSO. It now gets independent variant selection using the
select_variant selector clauses matching the loadable_module()
with output_name="libzircon", so it doesn't necessarily match the
variant used for the kernel or for the rest of userland, but it
will for simple and compatible variants like "gcc".

Original-Bug: 69697, 122877, 121817
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825536
Original-Revision: 5e539fce48864cd329e9f1801296dca1f09076c0
GitOrigin-RevId: de65d13917e690bc96295623266458cb035219fe
Change-Id: I0864fb5d30f8e801a526724a1f7cabc455a0509f