[roll] Roll fuchsia [network/tests] Improve socket buffer size checks
Verify that the SO_SNDBUF socket option is set correctly for the TCP
socket benchmarks before proceeding with the test in the same way that
the UDP tests verify the SO_RCVBUF. Failure to set the socket option
will result in dropped packets and the test hanging indefinitely.

Also add Linux-specific logic to expect the value returned by
getsockopt for SO_SNDBUF and SO_RCVBUF to be double the payload bytes,
since the man page states that these sockopt values are doubled when
set to account for overhead.

Original-Bug: 123899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832077
Original-Revision: 1ad252e0589b67abebc3d4fce53a3fb256fb4b28
GitOrigin-RevId: ee3c7ffcbe8c4f88a56ae053925c3e14a4eb8ffc
Change-Id: I3f8875da8aa272505bb1c1874a4bee2f4f3d1bf3