[roll] Roll fuchsia [msd-intel-gen] Fix initial race condition
The `wait_future` and `restart_future` might run out of order.
Create the test connection in the construction of the llambda instead
of the first time that it is run to ensure that it will always
be created before Restart is called.

Also set `connection_` to zero initial so the null checks in the
class will fire correctly.

Multiply: msd_intel_gen_integration_tests_component
Original-Bug: 125136
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833219
Original-Revision: 3bc4679e16e5861be07b1392ae0f4507a33538d9
GitOrigin-RevId: d6cea0f3fc35a8caee90a4126c774653a6f1f492
Change-Id: I12bba38a5b894a615e17a73178b8c7ea811b2927