[roll] Roll fuchsia [dfv2] Fix loader under ASAN
In DFv2 the loader is manually installed in the DriverHost. This means
it does not go through the normal process setup, and so DriverManager
must manually configure the loader for ASAN.

Also, the Compat driver should not be cloning the loader, as a clone of
the loader loses configuration information.  Instead, update the Compat
driver to use an unowned connection to the existing loader.

Now that the loader was updated to make blocking calls, the backing
loader in the tests are run on a separate thread.

Original-Fixed: 105603
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832781
Original-Revision: b4986fc37de2647bd11680437e0eae86fe7ae618
GitOrigin-RevId: f4189b1f097bf736c60c996d3418fed09503203f
Change-Id: If0909de99c59ff1bc7e3a9772a729266a1dc66da