[roll] Roll fuchsia [starnix] Fix SelectTest.SetrlimitCallNOFILE
This test failure had nothing to do with RLIMIT. Instead, we were
failing this test because select for no events on a tmpfs file caused
and error instead of blocking until the timeout. Previously, select was
implemented in terms of epoll, which produces an EPERM error in this
situation.

Original-Bug: 76815
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832736
Original-Revision: b065038711f386fbba6ba342126d97f363117196
GitOrigin-RevId: 4a68ed020710c804ccd0d7d1707d760eb442f1f4
Change-Id: I5208de590521438b3e5c0bd84b251542e38edc93