[roll] Roll fuchsia [amlogic-display] Explain primary node choice and "colocate" requirement.
DsiImpl banjo protocol is not supported by fragment-proxy for
inter-devhost forwarding. So the request for DsiImpl client
of DsiHost in amlogic-display driver may not success when dsi-dw
driver (implementing DsiImpl protocol) and amlogic-display driver
are not running on the same devhost.

As a result, the "dsi" node of amlogic-display must be the
primary node and amlogic-display must be colocated with its
parent. However, `colocate` property in its component manifest
is only an advisory so this may break one day.

This adds some comments and TODOs explaining the reason of
primary node choice and `colocate` requirement, and notes that
this is very brittle and can be subject to failure.

Original-Bug: 125227
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832749
Original-Revision: d0c27b3ea80eb70e1b718440851566e74261e82a
GitOrigin-RevId: 9593924e5e08338ec54b35be181e182051666a5c
Change-Id: I1d002114b05f9a81aa65ed81b4448aaa357a77f8