[roll] Roll fuchsia [syslog] Don't assume va_list is safe to use uninitialized
It's not really kosher to use va_list without va_start
intializing it. But in practice it's sufficient to explicitly
default-initialize it. It's never OK to just use an uninitialized
value, even just for passing it to be ignored elsewhere.

Original-Bug: 122875
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833318
Original-Revision: 31f85aaec3305b59220e12f9fee0dbf9a4119c9f
GitOrigin-RevId: 09e3c4eee4c662bfaf98dfe2ffbbc968a4949ee9
Change-Id: Ibde53040769d82724e92ad0cbe8de80b93355000