[roll] Roll fuchsia [zircon][display] Move mipi-dsi lib under display/.
The library at //zircon/system/ulib/mipi-dsi is only expected to be
useful to the display driver stack. This CL moves the library to
//src/graphics/display/lib/.

This CL replaces zx_library() with a combination of library_headers()
and source_set() in mipi-dsi/BUILD.gn. This gets the library off of the
allow list in build/BUILD.gn, and seems simpler than reasoning about
including a zx_library() target in our display stack.

Otherwise, the CL tries to leave things as they are.

Original-Bug: 36063
Test: fx test mipidsi-test (no code changes)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833525
Original-Revision: 55b03cbd1a74d90e3adaede10a141333266096d6
GitOrigin-RevId: 9d232cbe81dc923ab0f2f5f1d348d8fac47527a2
Change-Id: I8c943369b0531088c9414b551d6a70ea1dd3b76b