[roll] Roll fuchsia [ffx][emu] Resizing oversize Disk image files
The FVM or fxfs file that is part of the product bundle contains the blobs for
the image. The disk image file then needs to be extended to allow for the
minfs or fxfs file system to be created. If the original image file was larger
than the target size, the image file was unchanged, and the file system could
not be created, causing the emulator to not completely start Fuchsia.

This change checks the size of the image file, and if it is over the target
size, it is resized to 110% of the original size to allow for the
file system to be created. Longer term the product metadata in the
product bundle should be updated to reflect the contents of the FVM
so an appropriate target size can be specified.

Original-Bug: 102866
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831479
Original-Revision: 9abb8c80460dbb2389002b42d57e9bac80d38e07
GitOrigin-RevId: 8693bd91594c9de1ca2ceff5aa6601c04c09e75b
Change-Id: Idfb9c96df04ee53c4d87eaf6697a66f0b5555c37