[roll] Roll fuchsia [boringssl] s/BSDish/OpenSSL for license tag
BSDish is a reasonable stab at the problem given the... free-wheeling...
nature of the LICENSE file. However, BSDish isn't actually a license,
and getting a more structured level of license tagging is a priority.

rhilary@ 's at OSPO approved doc for what licenses are and are not
concurs that OpenSSL is a reasonable tag for this unreasonable license
text. See
https://fuchsia.dev/fuchsia-src/contribute/governance/policy/open-source-licensing-policies?hl=en#approved-licenses

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834897
Original-Revision: 5f4d42ef17b5922ee696162d2994a95ec3f40be8
GitOrigin-RevId: 12e26c335ce469a9d5215c84da15fda534c12e15
Change-Id: I23d66c7c6811e68ff83831e4443f6349bc178cf0