[roll] Roll fuchsia [netstack3] Synchronize IP device config access
This change updates the IP device layer to perform any work that depends
on the IP device configuration only while having access to the IP device
configuration. That is, all actions depending on IP device configuration
is now performed while holding the IP device configuration lock.

This is to prvent a race condition with interleaving reads/writes on the
IP device configuration. Previously, the following race was possible w/
Thread 1 enabling a device and Thread 2 disabling a device when the work
to enable/disable a device was not protected by the IP device
configuration lock:

    T1: set `ip_enabled` to `true`
    T2: set `ip_enabled` to `false`
    T2: perform work to disable IP device
    T1: perform work to enable IP device.

  Result: The device has `ip_enabled` set to `false` but the last
          work performed on the device was to enable it (e.g. protocols
          like MLD/IGMP may be trying to send packets).

Note that currently, it is impossible to observe this race in practice
because there exists a single global lock in bindings that synchronizes
all accesses to the netstack (`Ctx`). This change is to prepare us for
a world where that global lock no longer exists to enable concurrent
accesses on the netstack.

This change requires several context trait implementation trampolines
as a temporary measure until IP device protocol contexts are refactored
to be more strictly scoped to an IP device. See https://fxbug.dev/125261
for more details.

Note that this change removes the check in the send path to make sure
that the IP device is enabled before sending the packet to the device
since that check should be performed by the routing table and the
lock that protects the IP device's enabled configuration may already
be held in the send path. See https://fxbug.dev/125338 for more details.

Original-Bug: 108806, 125261, 125338
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831456
Original-Revision: 3b1cf16616da6af8e8928896929242ddfab97921
GitOrigin-RevId: 89144fa235bb014189270dc16c5703adfe5f36c1
Change-Id: Iff585503d350f922bef83dbc536a275084cbf445