[roll] Roll fuchsia [driver_manager] Don't reload Driver for firmware
The FirmwareLoader is able to get the driver struct out of the device
instead of re-loading it through the driver loader.

This simplifies the firmware loader class as it no longer needs to use
the coordinator. This means that the coordinator doesn't need to forward
declare the class.

TEST: ddk-firmware-test
Original-Bug: 125288
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833856
Original-Revision: 44ac1a9a50446c22c881b9daa913a5e86973e5c4
GitOrigin-RevId: 30f2cee913513c74860c03ea8e2a644b19b0c3a2
Change-Id: I2695bacf6166ec5ef23cc2b7361dbd111f922aef