[roll] Roll fuchsia [bedrock] Implement Remote for multishot::Sender
This makes it possible for Senders to be used over FIDL in programs,
and inserted into CloneDicts.

For example, a program can use this to connect to a FIDL protocol by
sending a server ends through the Sender to the serving Receiver.

Original-Bug: 122024

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833524
Original-Revision: a210ac1efe9daf912700b48547e35234e74e5b94
GitOrigin-RevId: 58cee47b3bd253ca335bb88da03ffb2229f0fdf4
Change-Id: Ibfebbc25eef04ff0c243f75243821e6c3c40f501