[roll] Roll fuchsia [amlogic-display] Move capture support check out of Vout.
As described in amlogic-display/README.md, amlogic display engine
can forward the VPP (blended OSD) output to the VOUT instance for
encoding and display, or forward it to the loopback interface for
readback to the memory.

Thus, whether a display engine supports capture is unrelated to the
type of Vout, so we should move it out of Vout.

Currently there's some issue on VIM3 where display capture tests
are not working; so we disabled capture support on VIM3 display
driver.

Original-Bug: 124947
Test: On VIM3 with [case 1] TS050 touchscreen [case 2] HDMI display, run
      $ fx shell display-test --capture
      capture checks are disabled.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834696
Original-Revision: e7189a3c8e4040cfe2b307955deab854f03f3a68
GitOrigin-RevId: d3055d1716380943c48d71740a81c726a2b28359
Change-Id: I36feab2d0fad03033ec2b31667a31e8b1d7dcfa6