[roll] Roll fuchsia [libc] Fix buffer overrun in test
The pthread_getname_np function's API permits it to write
to the whole buffer provided. It's not valid to assume that
it won't write as much as the size_t argument allows, even
if Zircon thread names are never that long.

The llvm-libc implementation of strlcpy writes NULs to the
entire buffer size given, which is suboptimal but valid for
it to do.

Original-Bug: 122880
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834436
Original-Revision: b45b8a0a185de49ecb74032a9b3fd7b30e9e4c30
GitOrigin-RevId: a26eb4c4138c06f810c30ff9fc3265f4bcd5a064
Change-Id: I68ce08423a1569fbd77c15d4cf06fbfd9631844b