[roll] Roll fuchsia [package-tool] Package Manifests Lists are relative to working dir
It turns out `pm publish` assumes package lists are relative to the
working directory, rather than the file. Scrutiny relies on this, so
we need to switch `package-tool repository publish` to be compatible.

We eventually want to migrate the package manifest list to json (or
some other structured format), so for future flexibility, this keeps
the `create-all-package-manifests-list.py` to merge lists rather than
reverting back to `cat.sh`. This should make it easier to migrate
to a different format in the future.

It also migrates scrutiny to use `create-all-package-manifests-list.py`
to make it easier to migrate formats in a single place.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832245
Original-Revision: 0c9fe66d0d2c903f7f736ee9f9c5b23f3c842da9
GitOrigin-RevId: ab92b047cce2a882e5805dddce6d47dd86435ad4
Change-Id: I30f0b961e11739776d7f83352813ef681a4ff2cd