[roll] Roll fuchsia [starnix] Fix flock_test
We now pass this whole test suite. The flock operation is better
modelled as an operation on the FileObject than the node itself. To pass
the last test case, we need to fail flock operations on O_PATH
FileObjects with EBADF. This factoring avoids the need to pass the
FileHandle to the FsNode, which is a bad code smell.

Original-Fixed: 79272
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835073
Original-Revision: 21fe3b7cf2d03871fbb159d60a68c44c12c131b0
GitOrigin-RevId: aed4fd2c34ecbf6752b09e66b29c1ce80169cbdb
Change-Id: Ib53ab3b52fab3dc2fc18dbab54d2abe72c6651c5