[roll] Roll fuchsia [assembly] `create-update` can optionally rewrite default repo
This changes `ffx assembly create-update` to add a
`--rewrite-default-repo` argument that edit's the update package's
`packages.json` to use a different repository hostname. This was then
used to change scrutiny to use this to switch all the update package
urls to `test.fuchsia.com`.

This begins a graceful transition for the security pkg test to migrate
to passing in the `update_domain` rather than passing in a test
configuration.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833516
Original-Revision: 58a67f6fe52c1309ca9da302d4e62fb189952988
GitOrigin-RevId: 6ff104c44a3e66f8a1eca89a2a2e7c013009a640
Change-Id: I86695c64d51317b3b73415bd2374c7f269c73c03