[roll] Roll fuchsia [kernel] Add`and` command to kernel shell, use in unittest script
This adds a new meta-command to the kernel shell: `and ...` does
the ... command only if the last command succeeded.

This is used in the kernel-unittests-boot-test script along with
the `-r` feature of `ut` for the suites that previously the
userland invoker ran repeatedly.  This gets the kernel shell
script version of the unit tests back up to par with the previous
way to run them via the standalone core-tests userland boot test.

This also factors out the GN logic into a couple of templates to
make it easier to define boot tests based on kernel shell scripts.

Original-Bug: 122154
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835359
Original-Revision: 2a9f4891810994d94b7f010ed6b676e12c2bdb8c
GitOrigin-RevId: d2f1e02369ea996014acd08ce1c248694df2cc6a
Change-Id: I92db1d1c902608d6929772d69825eaa8678d82b9