[roll] Roll fuchsia [ffx] Make isolate names correspond to test name
When debugging a self-test failure in Infra, we need to be able
to find the logs for the failed test, which is difficult when
the isolate directory names don't correspond to the test names.

The rule in this commit is not super-strict, but roughly it means
that a test with path '<component>::...::test_<test_name>' will
be found under '<component-test-name>'.

Original-Fixed: 125417
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835877
Original-Revision: 89fbb9ed95da72c4476f689d7b2e46e1b64a62a3
GitOrigin-RevId: 3dde45da3120d7cbcdcc02ab77c3b6cea99303b9
Change-Id: Ic68fe3a5e471430edee71fef65abeeba6b283c76