[roll] Roll fuchsia [sys_driver_cpp] Explicitly teardown in destructor
On ZirconConnection teardown there are a few things that need to happen
in order:
1. Disable the notification handler
2. Shutdown the loop (cancels remaining async tasks)
3. Destroy the MagmaSystemConnection (calls into the driver to tear it
   down).

Currently steps 2 and 3 are run by destructors, ordered based on their
order in the ZirconConnection. This is a bit fragile, so it's safer to
run them explicitly in the destructor.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834938
Original-Revision: dd7ce014c2d712447ea6e7ba00b516f50bbd85e4
GitOrigin-RevId: 6c0bcd76259d28efce001348b29174f949b09d31
Change-Id: Ic8e663e9e9257093e7db5f3470dc66575333e2f8