[roll] Roll fuchsia [bedrock] Add RestartableElf
RestartableElf is a proof of concept of a component-like object that
manages the execution of multiple Elf program instances over time.

It can be started and stopped, which recreates an Elf program from
a declaration (path to the executable) and an incoming Dict, which
contains capabilities provided to each program instance.

Tests demonstrate a FIDL client that reconnects when restarted,
and a restartable FIDL server that handles multiple connections.

Original-Bug: 122024

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825375
Original-Revision: 4ebab07780b1bfbfa6ee1da517c6aad1190a8192
GitOrigin-RevId: fbc1fe57ceefc1775be7d2ad9aec3dff66662aee
Change-Id: Ic25bb6f331d1fd9c77dc91552444d176705b6851