[roll] Roll fuchsia [kernel][riscv][mmu] Implement EnlargeOperation for Unmap ops
In the case of an unmap operation that fails to allocate memory while
splitting a large page table, the EnlargeOperation option signals
whether or not it's okay for the underlying code to fail by simply
dropping all of the mappings that large page covered. Now is parity with
the other arch mmu implementations.

Original-Bug: 122875
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835063
Original-Revision: 7fbe40542f142a2f058caa91802ca36f3a6e735a
GitOrigin-RevId: 05b252429a5c3bace5aaca73b44db01e1c9deb69
Change-Id: I86751e15a1999d68d14fe8e71c7fc24ecadf2299