[roll] Roll fuchsia [devfs] Remove classes from multiplexing allowlist
Remove all of the classes that don't depend on multiplexing.

While we're here, add the protocol id to the error message to
make it more obvious which class path relied on multiplexing.

This was verified using CQ only. It is possible that this change breaks
programs that don't have CQ coverage. In such cases, the
relevant entries should be re-added one at a time instead of
reverting this CL.

Original-Bug: 112484
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834857
Original-Revision: 492b93bb173bd5f4bbac43380a9d6caa5d14380a
GitOrigin-RevId: aaa86c483a1ffafae2f91ddfa927d2c36265f20e
Change-Id: I73e978a85bcb56622ba30f7830fc62c69449cffa