[roll] Roll fuchsia [mock-mmio-reg] Hides MockMmioReg behind interface
Currently, the library requires users to create an array of MockMmioReg
instances only for the purposes of passing the array to the
MockMmioRegRegion's constructor (as a ptr). The tests themselves don't
really ever use the array itself. Also, the MockMmioRegRegion exposes
the underlying *Reg instances by way of operator[].

One complication is that this library is used across //integration repo
edges, which cannot be updated atomically. This adds an intermittent
backwards-compatible version. Once all the uses outside fuchsia.git are
updated, the compatibility shim will be removed (and MockMmioReg hidden
in a private namespace).

Original-Bug: 122816
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818728
Original-Revision: 005a16015daafe765f9813bbcb612cff289694fc
GitOrigin-RevId: 056d0b1c5bc5b0a2866b17a403f555bfe96fcee1
Change-Id: Id0156e6b04d8daeb5abec36cde7fb9bac20ffe3c