[roll] Roll fuchsia [netstack3] Don't use `Mutex` in timers tests
This change drops the `test_late_cancel` test since we no longer
have a way to "pause" the timer dispatcher task after it receives
notification of a fired timer. Note that this change also removes
an opportunity for the race that the test checked for since we no
longer have an `.await` point after the timer dispatcher is woken
up to handle a timer.

Original-Bug: 122725, 125388
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835793
Original-Revision: 08e98a73f07c920e8caf6abadd249b5020537b11
GitOrigin-RevId: c68bde9537c400ce6090af8174201aa0ff8b25b0
Change-Id: I14e8c35d7941a13e4c1d74af3617cba13888f499