[roll] Roll fuchsia [fdf] Add async version of start hook
In order to provide better synchronization mechanisms for clients
issuing bind and rebind calls to drivers we must know when start has
effectively "completed". This happens asynchronously and not at the
return of of the existing start hook.

A new lifecycle version struct was added with a new alternative start
hook. If present it is preferred and chosen over the older v1.start.
Because we don't actually have any stability guanatees we can possibly
compress these back down to a single v1 struct with a single
asynchronous start function. If we make that decision, it will be done
in a future CL.

The BasicFactory class was updated to always call an asynchronous
version of the Start method. DriverBase has been updated to provide
default implementations of two Start methods. Either may be implemented
depending on whether the driver wishes to reply asynchronously or not.
This does mean that if a driver fails to implement either that the error
would not be caught until runtime, however this is likely and okay
tradeoff vs forking the DriverBase into a second class based on whether
or not an async or sync start is implemented.

A future CL will update the compat driver to complete start only after
the init hook is invoked on DFv1 drivers, providing the same semantics
as we have in DFv1.

Original-Fixed: 124912
Original-Bug: 125339
Test: fx test driver-host2-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833038
Original-Revision: 732fc29ccc0ae1971c92964d6c3dee08d10b0526
GitOrigin-RevId: f2a69c2a5ce416349ae0e836d4e723b3092ae46b
Change-Id: Ib5c4258f048128d3647a14fdd903a57e6b535f65