[roll] Roll fuchsia [kernel][riscv] skip a lock dep regression test on single cpu runs
Skip instead of failing the bug #84827 regression test if booting with a
single core system. Though it's not ideal the test cannot be run it's
also not ideal to fail all of the unit tests because of the single core
situation.

Original-Bug: 122875
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836825
Original-Revision: a922a9df901eddc0d4e42775a00db72b9666cc48
GitOrigin-RevId: 32033749b7ad482a7a670245ee9cac5ce92ed574
Change-Id: I7528d33aa6223b1cd9bb4e056482118713783cdc