[roll] Roll fuchsia [aml-canvas] Remove macros.
This CL removes the macros in aml-canvas, because the C++ style guide
mandates that macros are only used when absolutely necessary.

The logging macros are replaced with direct zxlogf() calls. The extra
functionality in the macros is no longer necessary, because zxlogf()
includes the file name and line number in logging output. While at it,
this CL also does minor logging improvements, such as replacing numeric
zx_status_t values with string representations, and clarifying the
wording of the logging messages.

The alignment-checking macro is replaced with a function, which now also
checks its precondition.

Test: fx test aml-canvas-test
Test: Booted core.vim3
Test: fx shell display-tool squares
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837218
Original-Revision: 44dafd0ee9caf13b21262facb2a2172265d0ca8f
GitOrigin-RevId: b78f6377c6778baf66b0d5041ab2a592320af33b
Change-Id: I6b50980ea82be334781c1e8c0e7da3fdb928e03c