[roll] Roll infra/recipes [recipe] Add "autogenerated" tag for Multiply shard messages
When a build contains "Multiply:" shards, the builder will send a gerrit
message on "create test shards" step.

Currently this is not tagged, so Gerrit will treat it as human comments
and will not fold it automatically [1]. This greatly reduces the SNR of
CL comments when there are multiple CQ retries.

This change adds a tag "autogenerated:fuchsia-build" which Gerrit UI
uses; by default, only the comments of the latest patchset will be
displayed for bot comments [2].

[1] Gerrit:
    https://cs.opensource.google/gerrit/gerrit/gerrit/+/master:polygerrit-ui/app/elements/change/gr-messages-list/gr-messages-list.ts;l=247;drc=1c3479b1bfa6b16f13c1c4c3b59cd8897a489d98
[2] Gerrit documentation:
    https://gerrit-review.googlesource.com/Documentation/rest-api-changes.html#review-input
[3] Prior arts for Jenkins: https://github.com/jenkinsci/gerrit-trigger-plugin/pull/366

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/837136
Original-Revision: 67b604b49a79ee0e799edab166744d1b7addd332
GitOrigin-RevId: 9f57699ada674ce601c2f23d90c93ab865984f04
Change-Id: Ibc58a83de3983dc8d76c48d6ce54f3e04ee0aff8