[roll] Roll fuchsia [async-utils] Support keeping hanging-get server hydrated
...with an in-flight request.

This change updates `async_utils::hanging_get::client::HangingGetStream`
to be configured to eagerly send a (hanging-get) request so that the
server is able to serve a response as quickly as possible so the
`Future` is able to do meaningful work as soon as possible.

Before this change, a request was sent during a poll following:
  1) the creation of the `HangingGetStream`
  2) the completion of a previous in-flight request/`Future`

Now, a request is sent immediately after the above situations (before
polling) if configured to do so.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837056
Original-Revision: 156362a2185945ffedc3d1d009383755d935bd0d
GitOrigin-RevId: 900228968be828b12766615f216a398fe8506d76
Change-Id: I50dde28e6bc9f6e0a3a2c7cd09910a9d427e8275