[roll] Roll fuchsia [netcfg] Hold DHCPv4 client-ends in a `StreamMap`
This change introduces the first-steps in implementing the client-end
of the `WatchConfiguration` hanging-get. NetCfg how holds a stream of
DHCPv4 configurations in a `StreamMap` and eagerly sends a
`WatchConfiguration` request to the DHCPv4 client component.

Note that this change does not yet handle a response to the
`WatchConfiguration` request; that will come in a follow-up.

Original-Bug: 125247
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837077
Original-Revision: ed1c530ff284f577e71163fce110ebf125fb1f28
GitOrigin-RevId: 023b8766dab143dced364866b794fe7cd52f5b08
Change-Id: Ide8b125546e402a48aeb9e6b1ca864a08465a727