[roll] Roll fuchsia [magma] Split logging initialization into PlatformLoggingProvider
Initializing logging requires a dependency on PlatformHandle, which has
a dependency on several other libraries. It's also not necessary at all
with the DDK logger, so it makes sense to split it out into a separate
library that's only depended on when necessary.

Original-Bug: 111471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833339
Original-Revision: 6beee4396ef56ae5504916e5a13d29cffbc075e1
GitOrigin-RevId: 5e2d088aa254f70536ddb59c40bfad11b3e0c298
Change-Id: I948db48c2cd7b758daadef677c10887f64767d23