[roll] Roll fuchsia [compat] Serialize device_unbind_reply
Currently this executes on the thread the driver calls into the
libdriver API on, which will make thread-unsafe objects such as
device_server_ unhappy and assert. Scheduling the real work to occur on
the executor ensures it occurs on the appropriate dispatcher context.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838218
Original-Revision: 2f934802359795c87dbef8e0ab34f8e7d7d63250
GitOrigin-RevId: c67fc937a2577a3bbe84532cef72bb8d72f41524
Change-Id: I29d73c77747c802ab6dc64773855c22a4edafd4b