[roll] Roll fuchsia [fxfs] Fixed a flush/attributes race
There was a small window where it was possible to get old attributes
whilst a flush was taking place.

The fix is to check the in-memory attributes before retrieving the
attributes from the device (rather than the other way round which is
what opens up the window).

Original-Fixed: 125544
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838716
Original-Revision: c8b25a79b871475212ecf1cc95fbe9d1cc350aec
GitOrigin-RevId: f27c33327ef3779f4c78c9f1642f0abe50288f0a
Change-Id: Id5aa4cbc29a5cf9e8d6efbfb3cce6334518975fe