[roll] Roll fuchsia [botanist] Write blobs into tempfile in same directory as destination
Previously we would write files into the system temp directory, which
might not be on the same file system as the botanist blob cache. This
changes botanist to create a temp directory in the same directory as
the destination blob to make renaming the blob to be atomic.

This has the risk that this could leave behind tempfiles in the cache
directory if the go process is SIGKILL-ed, but should be cleaned up
with SIGINT.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838498
Original-Revision: 868f400ed0f32c8dabfd0020fd7dd1f39372e64f
GitOrigin-RevId: ec4905acda7eda8a537cc586821b2a103a233e85
Change-Id: I2414d083143411a0582d325fd81928a74dd333cb