[roll] Roll fuchsia [kernel][vm] Retain AwaitingCleanLength if possible
When a zero interval is split in the middle or clipped from the end, its
AwaitingCleanLength still holds. Similarly when the interval is split at
the start, its AwaitingCleanLength (minus a page) should be carried over
to the new start on the right. This ensures the AwaitingCleanLength is
retained across all kinds of splits. The same holds for when an interval
is clipped at the start, or when slots are populated starting at the
interval start.

Conversely, when unused interval slots are returned via
ReturnIntervalSlot, their AwaitingCleanLength should be merged back into
the resultant interval.

The only case we cannot retain AwaitingCleanLength is when an interval
is extended / merged on the left, because the AwaitingCleanLength only
describes a range that begins at the interval start.

Test: vmpl_unittest
Original-Bug: 122842
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837216
Original-Revision: 61bbddb32ffa24978eb4812797fbf9e9e1bd2ae9
GitOrigin-RevId: dc52bae9a850077e7cbfa4b41140662a91f7b4d3
Change-Id: Ib82ad906a02dbd9d1df030727209f087c86cfe88