[roll] Roll fuchsia [kernel][scheduler] Remove GetTargetPreemptionTime
Remove the GetTargetPreemptionTime method from the Scheduler.  It is currently
unused.  If similar functionality is needed at some point in the future, the
method can be re-introduced in a way which is compatible with the
thread-lock-free future which is fast approaching us.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839036
Original-Revision: 787add910bf66f5dff7debe3d30307f2003562cd
GitOrigin-RevId: 0843579b23f947d7f288f9237863cd2174efbb6f
Change-Id: I9f6e350749d7ab94a4d55955015cb69ec01312b0