[roll] Roll fuchsia [driver_runtime] Allow multiple dispatchers per thread pool.
The coordinator checks whether a thread pool with the same
scheduler role already exists.

A thread wakeup handler has also been added to the thread
pool's async loop, which will check whether the scheduler
role has been set for the current thread. This is as we
can't post a task to a specific thread in the pool.

Original-Bug: 124644,85946

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837857
Original-Revision: 8c50050c79dbed4433572d676df8615db8a17409
GitOrigin-RevId: 825e062e4d79b9ad6decd0fcf6e30a092a545c80
Change-Id: I36a3d26a0a1953ae75b0867d862798a445337875