[roll] Roll fuchsia [fx] Disable report-bug inside fuchsia.git
This change modifies `report-bug` to warn that it is no longer usable
and exit. Components are no longer specified in OWNERS files after the
submission of
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/834141

This file is an exceptional case of a haunted graveyard and is
essentially impossible to modify by any standard software engineer.
Because of this, it needs to be considered legacy code.

Analytics shows it is in single digit usage of the past month.

Original-Bug: b/278592638
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838056
Original-Revision: 954eecad2ca396a6b887970d16ed7e5d812b6370
GitOrigin-RevId: 5e96a47ce863780b74297f83a747ef4a24358b33
Change-Id: Icf952a903024775450fdbf4f49a4a24f1ef13653