[roll] Roll fuchsia [starnix] Add a test for mprotect failure.
This CL adds a syscall test that makes sure starnix "doesn't get
confused" if `mprotect` fails. The idea is that `mprotect` could fail on
any part of the provided address range, and it is not defined what
should happen to the other parts of the range.

The test make sure that only two options can happen: either the mprotect
change succeeded or it failed. In both cases, it also tries to access
the memory regions to make sure that the actual mapping rights
corresponds with the reported ones.

Test: fx test syscalls_test --test-filter "*MprotectFailureIsConsistent*"
Original-Bug: 123205
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824390
Original-Revision: 3cfe683da2790344730bbbd2879507fe6f2b75b4
GitOrigin-RevId: acd2c511d7b3e485be7c7b20bbd6905a0abcff09
Change-Id: I85bc12e9d27453558cb760ea2977e964d17ad41e