[roll] Roll fuchsia [ffx] Only do extreme build configs for integrated ffx
When building ffx for separate subtools the binary is much smaller and
we can relax a bit in terms of our optimization choices. This changes
the ffx template to allow for this and uses the separate subtools flag
to choose the value of that flag.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836156
Original-Revision: bf6950a95387f8bc85da5f8535d5cfe26e9af1ef
GitOrigin-RevId: 152367abf13300c3e568c8636ed7151e94d7b1f9
Change-Id: I26a4f4694f1c93495f675f8247a52f4c4e7d2f82