[roll] Roll fuchsia [check-licenses] Updates to file package.
In the future, license analysis will run in multiple threads,
so each package needs to be concurrency-safe.

This CL un-exports all public fields in the file package, opting
instead to use getters and setters everywhere. This is necessary
to prevent race conditions.

It also introduces custom JSON marshalling logic, since fields need
to be exported in order to be read by the default json marshaller.

Finally, it generates the SPDX IDs for files and filedata objects
by using a hash of the data. This should result in more deterministic
output files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838096
Original-Revision: 435183f396cd60424665183aa1658660e9ac2cc1
GitOrigin-RevId: 6be86f2722729d1897f7759169d7164e3875f9c8
Change-Id: Ia07ef9ec57911d9cb4226df14f54687bc1657257