[roll] Roll fuchsia [fshost] Handle zxcrypt wrapping inside ZxcryptDevice.
This removes the need for plumbing of inside_zxcrypt and special casing.

There are still a few asymmetries here that are a bit ugly IMHO, but
I'll leave those as an exercise for the future. For instance, it
would be nice if SetPartitionMaxBytes was a Volume protocol call rather
than VolumeManager as this would allow zxcrypt devices to transparently
account for their own overheads and further simplify code.

Original-Bug: 123427
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837237
Original-Revision: 4969ef4711991c31c3b43eb580e5054c02baec63
GitOrigin-RevId: 19988bd7ac0178e86a7b1e9c4404b426cc0db164
Change-Id: I555d709d2d563e98ef9965bea20cf4a2820524a0