[roll] Roll fuchsia Reland "[starnix] Implement POLLRDNORM and POLLWRNORM"
This is a reland of commit 11818c056255945ba78dc99fb66c55101ca9cf4b

This CL is the same as the previous landing, except with the flaky test skipped. This test is flaky because it's using a race condition to trigger an interesting error condition.  When the error condition doesn't trigger, the test passes.  When the error condition triggers, we fail the test because we don't implement the error condition correctly yet.

Previously, we always failed the test because we didn't behave correctly when the error did not occur.

Original change's description:
> [starnix] Implement POLLRDNORM and POLLWRNORM
>
> We need to set these FdEvents as well as POLLIN and POLLOUT.
>
> Original-Fixed: 79315
> Original-Reviewed-on:
https://fuchsia-review.googlesource.com/c/fuchsia/+/836416

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839678
Original-Revision: ac34b912221d8f5d632bea544fcc0592430508f8
GitOrigin-RevId: 24235bd7da13a1b3516a2d1995d83b458a4565c5
Change-Id: I59f5545dc73c2d5791fed1548dc78ec130fda311