[roll] Roll fuchsia [starnix] Request backtrace in CoreDump crashes in restricted mode
When the restricted mode executor is about to exit a task due to the
taks state being set to a CoreDump exit code, solicit a backtrace from
the system by generating a "backtrace request" in the form of a software
breakpoint exception with a magic value in rax / x0. This also (re)emits
CFI directives to help unwinders see into the Linux state at the point
of exit. This is particularly useful for Linux abort() implementations
that raise the SIGABRT signal on themselves.

Crashsvc's inspector library cannot correctly unwind this stack yet, but
it should be able to soon.

Original-Bug: 124427
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838478
Original-Revision: d4a608600362edb845247328e2a4b18334dcd255
GitOrigin-RevId: 1ca9b0cbcdc62ef8101e956b1f3c0a21f50a7583
Change-Id: I4f4f410bb73a094f53d1f8c69dd9ef8d3cd11f08