[roll] Roll fuchsia [compat][test] Add StartDriverArgs
Update tests so they pass configuration in via a struct instead
of function parameters. This makes it easier to have defaults
for tests that don't want to change behavior.

Update the TestFile object so it can only be configurated on
creation.

Both of these changes are a setup to let the test infrastructure
run on a separate thread.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/840056
Original-Revision: 0f7fa2fd2f4275ccf6808ea98df997cd3da1fdc7
GitOrigin-RevId: 40f106d326afcc388d13aea556593ece2acfae75
Change-Id: I93eafb584be4ef917b71c90db59c02177f240202