[roll] Roll fuchsia [compat][test] Add IncomingNamespace class
Pull all of the mocks out into their own class. This should not change
the threading behavior, but it makes it possible to put this class into
a DispatcherBound object in a followup CL.

The IncomingNamespace class is entirely configured when it is started,
so TestProfileProvider must be updated to also be configured on
startup.

Multiply: compat_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/840057
Original-Revision: 534dc43828b36267c296815e6ccf52ae9cee7f96
GitOrigin-RevId: 2a0f849a61c9da31c1d6d4291bad012e1c79d103
Change-Id: Iec1bef086cfdb1cc2c14aad4c820df0633908bb6