[roll] Roll fuchsia [wlan] Modify softmac beaconing APIs
Removes ConfigureBeaconing() to disallow updates to beaconing once it is
started. Right now the API surface only provides an EnableBeaconing()
and a DisableBeaconing() to keep it simple. When there is a need to
update beaconing, the suggestion is to create a new method named
UpdateBeaconing() or such, and pass it only the set of parameters that
we would like to allow in an update method.

Original-Bug: 87050

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831321
Original-Revision: 0193f195eff6c62ef60f37976837facc9e61357e
GitOrigin-RevId: 507420ed3fd13c1fd10bf807725246b8ce8a5c0f
Change-Id: I1a3614470977492e1feacca561452c3ba07d7532