[roll] Roll fuchsia [starnix] Delete Credentials::from_passwd
from_passwd ignores almost everything you give it, which is confusing
and makes it seem like the username and password specifications are
doing something when they aren't.

For containers, it seems simplest to get rid of the init_user
specification. Everyone is just setting it to root.

For components, introduce a "uid" key. Let the gid default to the same
as the uid, as that's what every single current use of "user" does.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839458
Original-Revision: 00f54040fa24de17ad69397ab61f3e582f71c65e
GitOrigin-RevId: a96812c1df8594ab89b3ebe6952f81964abc7da2
Change-Id: I64bf70f5046b0e3c34535f29254c413bc8b50fa7