[roll] Roll fuchsia [pci] Configure Bridge IO windows before device BARs
Ensure bridges get first priority when it comes to allocating address
space to prevent cases where we need to reallocate a device BAR earlier
in the topology and end up allocating space that firmware has already
allocated to a bridge window later in the bus. This can happen on NUC11
with the bridge upstream of the ethernet controller.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839739
Original-Revision: 4313010bf35db2c72312c2f976879199b91f75f0
GitOrigin-RevId: b177e6b23525236852b8781802855c13b0888878
Change-Id: I19ac1af24f65f104fcab641a62a56b0ecba36013