[roll] Roll fuchsia [input] Generate a distinct event_time for InputEvents during binding
Previously the event_time from the InputReport used to create an InputEvent is being reported as the event_time for InputEvents. Instead, device bindings should report the time that an InputEvent was generated in the event_time field. This is needed to write last_generated_timestamp_ns to Inspect.

Original-Fixed: 125731
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/840356
Original-Revision: 6ec02895f267c86a782dcbd8ba11c173813b24b0
GitOrigin-RevId: 6bd347bf13a55bb85cbf7e735372fd494dabcb78
Change-Id: I3c05dd14702ad8869ab5c3f4ce8b9bbcb3add8c2