[roll] Roll fuchsia [dfv2_v1_lifecycle] Don't use DevfsExporter
Use the new node devfs API instead of DevfsExporter.

This requires adding a "dfv2" node in the topology which is awkward, but
this code is in the process of being re-worked and there are no real
clients for DFv2-in-DFv1 at the moment.

This is the last user of DevfsExporter, so it can now be deleted!

Original-Bug: 125288
Original-Bug: 117760
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839063
Original-Revision: cac7585943936541e710d36ceb9e6fb40a22b11e
GitOrigin-RevId: 1dbe753ef73624df7036607c827b053470b7b536
Change-Id: I82c2bd5536e169ebee37189dbd16109d3d8f41f6