[roll] Roll fuchsia [amlogic-display] Use 16x16 SPLIT_BLOCK formats for AFBC by default.
Upstream amlogic-display driver requires SPLIT_BLOCK modifier bit
set for AFBC images with 16x16 block, or 4k AFBC images with 32x8
block. If SPLIT_BLOCK is not specified, for large images (e.g.
width = 1920), there'll be blocks / lines of pixels displayed as
random colors.

Besides for better performance, we also added YUV and TILED_HEADER
modifiers; these format modifiers are not used in upstream amlogic
driver but we have verified it's working for vulkan applications.

A side effect of this change is that it breaks display-test, as
the encoding logic of display-test is not compatible with the
TILED_HEADER and YUV modifier we added to the display-driver.
fxbug.dev/125730 tracks the effort to re-enable the test for
amlogic-display.

Test: vkcube-on-fb on astro, no regression (afbc enabled)
Test: vkcube-on-fb on vim3, no regression (afbc disabled)
Test: vkcube-on-fb on vim3, after patching fxrev.dev/838618,
      afbc can be enabled
Original-Bug: 125730
Original-Bug: 124949
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838617
Original-Revision: 9512be8b036d6552e4bc7935aa15d71a0adb7c57
GitOrigin-RevId: 16b5af25e5f3f61864bfdec3e571b20fa115f696
Change-Id: Ib217cbd11e6a0eb7ff16399aad1c528d75e05d18