[roll] Roll fuchsia [sdk][ffx] De-dup ffx_tool atom files entries
Based on feedback from the initial sdk metadata, change the files
section to be an object with three fields:

- executable: the binary, previously at the top level
- executable_metadata: the fho metadata file for the executable
- support (optional): other files the subtool uses

Previously the executable* items were at the top level, and then
duplicated in files. This was somewhat convenient for writing
distribution logic for the tools, since it's the same as host tools, but
this makes things a bit clearer and we can make the logic down the line
handle it.

This also updates the build and ffx sdk tests to correct for changes in the
metadata formats.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837997
Original-Revision: f3b8e5fa5f05203c78d39e96571dd113890375a5
GitOrigin-RevId: fadda5e1e504cfc1fecb73957141a9f776ab4e1e
Change-Id: Iacb4b0bfc4cdb47565ac3a7b7ef6ab886b2053b5