[roll] Roll fuchsia [zx] Make zx.signals private
`zx.signals` should not be modeled as a mere alias of uint32. It is a
`bits`, but more thought needs to be given to model the current
public/private split of the current C definition; until then, keep the
FIDL definition private (and, while we're at it, capitalize it per the
FIDL style guide).

There are no other uses of zx.signals outside of fuchsia.git.

Original-Bug: 109734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837917
Original-Revision: b13c53018960eabb49a6ed70a3a72d54f2f6cd77
GitOrigin-RevId: 878cc6dbd3789cfc5a7f9859794ea0279d45db1c
Change-Id: I84795545b5b8a12b2e4fb16217c78e81847691a2