[roll] Roll fuchsia [audio] Include configs in pipeline-benchmark packages
In the vast majority of cases, tests do not need to modify the system
image simply to provide a config file to the code under test.  The
pipeline benchmarks appear to currently rely on passing configs through
the system image; this patch causes the tests to resolve the configs
from the test package itself instead, which is likely better for
developer iteration speed as well as simplifying product assembly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841963
Original-Revision: 62fffb4642cebcc63331e4bd7075c6d45b987ddf
GitOrigin-RevId: 96580ee0f7a9e9537a5e3c2203a24e4fb71b3c26
Change-Id: I26bc7dab980b2830dbff7bb1d92794e1ae2fcc6e