[roll] Roll fuchsia [dfv2] Split Runner into separate class
There is an API boundary for loading the driver components through
the component runner API. Pull this code out into its own class so it
can maintain its own state and be encapsulated.

The main reason for doing this is because DFv1 wants to start loading
DFv1 drivers through the driver runner. This is much easier to do when
the driver runner is unaware of DFv2 objects like Nodes. This will be
handled in a followup CL.

Original-Bug: 125288
TEST: driver-runner-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836756
Original-Revision: 02617383dd813621777db3fc72128ff2c8f06327
GitOrigin-RevId: a9439c0357f299ded9ee707662835411181bc747
Change-Id: Idd3dcc8ca6fbabc0a599848fa05937138d0b6ef9