[roll] Roll fuchsia [fdf][ti-tca6408a] Forward metadata
Forward the GPIO pins metadata when creating the ti-tca6408a device so
that the gpio driver can access it when the driver binds to the device.

Originally in dfv1, the gpio driver was still able to get the metadata
because in dfv1 the driver manager searches the entire device tree for
the metadata. In dfv2 however, the metadata is searched for only in the
parent device causing the search to fail.

Original-Bug: 125848
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841556
Original-Revision: 0e96873976884bbe704228bc47421669e9f6e43c
GitOrigin-RevId: 8f94cc124e1f4a1112491dcaf72f5ccf640834dc
Change-Id: I48e3fdbc9dc82a38359780f933942d7a9d6b7a6e