[roll] Roll fuchsia [kernel] Update 'error code' in r-mode tests
These tests were returning an 'error code', which is somewhat ambiguous
in the context of these tests. Instead we rename this to
'reason code' to match the terminology used in the restricted mode
documentation, and also update the vector code to populate this with the
exception code provided by zircon.

Original-Bug: 121512
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830782
Original-Revision: 83d7ff4cb82904c1fef59dc3e162161bda09d9d3
GitOrigin-RevId: 49e0e82a31112a3acb63c7304bd39d32a84ccddd
Change-Id: Ibcd3e2253d551f41fb04195d0b6785f48cf9a007