[roll] Roll infra/recipes [ninja] Remove -Werror from rpmalloc compilation.
The Mac build of the Ninja recipes is failing again because,
this time with:

```
error: unknown warning option '-Wno-unsafe-buffer-usage'; did you mean '-Wno-frame-larger-than='? [-Werror,-Wunknown-warning-option]
```

This CL modifies the way rpmalloc is configured to ensure that
it no longer uses `-Werror` when compiling the library. This has
been a source of constant random breakages in the past that we
tried to fix by just adding / removing specific `-Wno-xxx` flags,
but this approach does not work.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/842319
Original-Revision: 9a2899802cbc3231d18ae121c48e2469b0da60f6
GitOrigin-RevId: 800ba527f19465b857735995d7b64bc553261464
Change-Id: Ic60ba692290d5b65ae343816c3850dbab6658c24