[roll] Roll fuchsia [build] Factor hermetic_code_blob() out of code_patching_hermetic_alternative()
This refactors the machinery used for hermetic leaf alternatives
in the kernel code-patching support into a new general-purpose
hermetic_code_blob() template. This can be used for all cases of
hermetic code blob linking, not just the code-patching case. It
optionally supports a code blob with .rodata as well as .text
sections, which is not allowed in the code-patching case. It also
has machinery to make it convenient to share a local API header
between the hermetic code and other code that uses the blob.

Original-Bug: 125717
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841116
Original-Revision: 2afa0720d3bf9d9ddf76333f675018d669960728
GitOrigin-RevId: 1145290460200625c6ead4c93f4228d93b26fc1a
Change-Id: I2e74d02a407753e60b3a5953a884fe77b24ac32c