[roll] Roll fuchsia [blobfs] skip blobs that aren't written in health check
If a blob has size zero in the health check, skip it. This is either
because it's the null blob, or because it's not in the kReadable state.
Either way it's not super useful to health-check it.

This hardens the health check so it can be run while blobs are being
written.

Original-Bug: 124968
Test: fx test blobfs-integration.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842856
Original-Revision: f2d631a9d4e7a05edad4c7d1b0e04f6e5bbed7c2
GitOrigin-RevId: de0874ceaa00adce3d164742d794466337a40423
Change-Id: Iab34bd5466eba8bc426cb2b80aa7af78a3e5b7c4