[roll] Roll fuchsia [rfc] Clarify scope of RFC-0154 and RFC-0208.
The emphasis on the testing-related motivating use cases can lead the
reader to the impression that these mechanisms are _only_ intended to
be used in testing scenarios. I've added a paragraph to each,
indicating that production use cases are in scope, but there will be a
higher quality bar to use these in prod.

I'm sending this amendment around to many of the original
authors/reviewers to confirm that this was the RFC's original intent
(or an acceptable extension thereof).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831556
Original-Revision: f025258e00cfc6449f0bb0ee0a443da430391a33
GitOrigin-RevId: daae9dd4398aeb733423145c6d48686f4992d055
Change-Id: I5841cc01e1a41673890e6da712743ae910063276