[roll] Roll fuchsia [log][rust] Allow the initialization library to work on the host
This change also implements an Event Formatter which is basically the
Compact formatter with support for tags. The goal is to support well the
fuchsia crate which allows clients to set tags and are currently ignored
when using the default compact formatter.

Original-Fixed: 110026

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842278
Original-Revision: 5a602a907fafdc232d64afd2be9557b92b17f086
GitOrigin-RevId: 83adadf78f7d84e4feda773e15266f2695a450b3
Change-Id: Id4a51b5a06f5f22812baa39692274cb25af7075a