[roll] Roll fuchsia [ulib][test] Incr boot-libc-unittests t/o on riscv
This bumps the timeout from 600s to 1200s for boot-libc-unittests
running on riscv. A TODO is added to remove the special timeout
handling.

The reason for this change is to enable `bringup.riscv64-debug` builders
to pass when running boot-libc-unittests.

Original-Bug: 125957
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843136
Original-Revision: 1b14f74d07092bcc557bd0de130152f61dad2332
GitOrigin-RevId: 0d45c788d733dd67222dae95468a050f6986a78c
Change-Id: Ie40e293e0097dc70ee6d7807c00e428345d5f3aa