[roll] Roll fuchsia [ffx][fho] Properly support legacy machine output
The previous implementation of all of this didn't properly support
machine output from legacy plugins. It turns out that fixing this
required a fairly deep refactoring of fho because of assumptions about
how machine output state can be determined.

Many of those problems came from severing the `FfxTool` and `FfxMain`
traits from each other, which was done to make it so that FhoEnvironment
could be stored properly between them. We made that not necessary by
switching to using an Arc<dyn Injector> which let FhoEnvironment become
Clone, so this separation wasn't needed anymore and is undone here.

We also move building the 'main' object to happen earlier, simplifying a
bunch of things about the lifecycle of the object.

And last but not least, we break the `ToolRunner` implementation into
two separate ones, one that implements the tool and one that prints
metadata, rather than doing the switch between them late and awkwardly.

Original-Fixed: 125869
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841458
Original-Revision: 5d75532914c885f72bc2a74fc485310d14a0537b
GitOrigin-RevId: b29b503240af6b964a07233a9a85292610d83625
Change-Id: I09b3d216c8d6ce3a58c986df0c3aefc8af104281