[roll] Roll fuchsia [dhcp] SubnetMask type uses net_types PrefixLength
The DHCP server configuration defines a SubnetMask type that overlaps
substantially in functionality with the (later-introduced)
net_types::ip::PrefixLength type. This change updates SubnetMask to wrap
PrefixLength and delegate all overlapping logic to the latter, but
preserves the existence of SubnetMask in order to be able to implement
serde and FIDL conversions on it.

The serde formats are preserved by deriving serde::{Deserialize,
Serialize} on the old version of the type and then manually implementing
them on the new version to forward to the old version.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843897
Original-Revision: 174d9dcba0874af65d03395635cabf9010dac2a2
GitOrigin-RevId: 97b22213ad0704d7b18e8e58f2478c9d54d2acbc
Change-Id: I54c583bb71715cdb63f7e8a2fd8a1afbfb5ad05f