[roll] Roll fuchsia [dhcp/client] Bundle component with Netstack3
Route the DHCP client provider capability from the netstack component to
netcfg, and include the still-under-construction DHCP client component
in the Netstack3 package used for production.

Routing the capability from the netstack package itself makes the choice
of whether to use the built-in or external DHCP client configurable
per-netstack instead of requiring an orthogonal flag in netcfg or a
setting in the network realm.

Netstack2 does not expose the capability, and so will continue using its
built-in DHCP client, while Netstack3 will use the new component.
Because of this, scrutiny verification fails. Since the RFC-0155
mechanism for supporting this use case isn't implemented, add an entry
to the routes exception allowlist.

Original-Bug: 125773
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/840519
Original-Revision: 72815bc1117411d9f3c5c31683c62a63a9fa7b24
GitOrigin-RevId: c2bc8b85ded3409f7c95d997a0a9a8f9758df22e
Change-Id: Ie4d4db77dfdd43a8663f55a20e9af4e712ef97aa