[roll] Roll fuchsia [cobalt] Use test config data from package, not system
Tests changing the underlying system image by including config_data that
are only consumed by the test component are more of a pain to iterate on
(and for product assembly to construct system images for) than properly
hermetic tests.

This moves the config_data resources into the package, and then updates
the cobalt testapp to look in the test package for that config data,
rather than in a system-global location.  Since it appears the tests
were otherwise hermetic, we can also take them out of the system-tests
realm.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842116
Original-Revision: 4150058b1b6d0bc06eec1a19ad6fb3f58e7ec742
GitOrigin-RevId: 8ce78b6229aeed331713120ebe74de1bef79d2d7
Change-Id: Idd512225f5da5c5a5c8c4cc70d919756bea79506