[roll] Roll fuchsia [zither] Introduce 'output namespace' knob
"Output namespace" is defined to be the namespace/layout under which
backend outputs are generated within the specified output directory. By
default, this is backend-specific. The value can determine the name of
the resulting library/package/crate/etc. (when a function of source
layout), as well as the 'include' namespace of the headers generated by
a C family backend.

For now, we just add this support for the C family backends (though this
support can be trivially extended to the other data layout backends when
a use-case presents itself). An explicit override will permit us to
remove the header guard rewrite logic currently hard-coded for the "zbi"
and "zx" FIDL libraries.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828896
Original-Revision: 1bad414b5110eb83f12d567ca1f05dfa1e792bf6
GitOrigin-RevId: c45e30095282f54fb7b90da2a1dd46f80d00ddfb
Change-Id: I71133d55e190962af2975a35b7d2758156974005