[roll] Roll fuchsia [fxtest] Add helpful flag suggestions.
Many people have missed the '--output' option to print all output to
stdout. This change ensures that the configured behavior of output is
reflected in information at the top of the invocation, and it provides
suggestions on how to change that configuration if desired.

Original-Fixed: 126087
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/844120
Original-Revision: d1802f7d75e8503607fefaa3911b3dc7b26ec352
GitOrigin-RevId: a5dc102724d69589e4f7acefd73c04d374768363
Change-Id: I2c867c71479ca4f45a75904978777f31a7eed503