[roll] Roll fuchsia [bt][obex] Define a type for OBEX request packets
The OBEX Request (e.g. Connect, Disconnect, ..) contains an OpCode and
potential headers.

Implement Encoding and Decoding for this type.

Define the HeaderSet as a collection of Headers that can be sent with an
OBEX request or response. This collection is initially simple with a
TODO to improve and enforce ordering properties defined in the OBEX
specification.

Original-Bug: 120012
Test: fx test bt-obex-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832208
Original-Revision: 326fb5fd305eabc6dfa9fa660bf1c8a61ddfbdf0
GitOrigin-RevId: eafaf1191280d7e9013f94223593b2f10ec5adae
Change-Id: Ibb1d1ddc8e88f3271182ff7c78f1d4dd3000116f