[roll] Roll fuchsia [kernel][riscv] Fix bug in Map's returned count
The Map function currently does not include existing entries in the
count of the number of pages mapped. This contradicts the behavior
documented in
https://cs.opensource.google/fuchsia/fuchsia/+/main:zircon/kernel/vm/include/vm/arch_vm_aspace.h;l=72
and causes the failure in the attached bug. This CL adds the existing
pages back to the count.

Original-Fixed: 125951
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842719
Original-Revision: 732349c0f587a3c0c1c3fb9af45c7ea426319a8d
GitOrigin-RevId: d4a043cf8f3528360b1f9c8f363fa06fa7321ef7
Change-Id: I50d0b0341426124f2570cb1e8238d85e8b087ad9