[roll] Roll fuchsia [rbe] Support remote != local command
Distinguish between remote-only and local-only commands.
Accept both, while still defaulting to remote == local commands.

Sometimes the command for remote execution is modified so that
it cannot run locally, but it is expected to produce the same result.
For example, with remote-cross-compiling, the remote request's command
might use a binary for the remote execution platform, which is not
suitable for local host execution.

Special handling for local fallback: when the remote and local commands
differ and rewrapper failed in one of the local execution modes,
retry with the local command we were given if it is different from the
remote one we gave to rewrapper.

Not applying this to racing mode because rewrapper will ignore local
failures if the remote one succeed.

--compare mode will use the correct local command when it is different
than the remote command.  This helps verify that different commands
indeed produce the same results.

When local and remote execution results differ AND the commands differ,
show the unified diff of those commands.

Original-Bug: 105799

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845016
Original-Revision: 38a079fcbcdc99e35e8537bf5d95fdbcb7d988ca
GitOrigin-RevId: a919eba7ac2bf4bd9ef611a3ca4227d0cf741d81
Change-Id: Ifdca07f37f823c0c7fe1deb81619d7cefd0294d2