[roll] Roll fuchsia [ffx][fho] Only output any given tool once
If a tool exists in both the sdk and workspace, or the ffx binary itself
and either of those, it'll currently print twice. Use a hashset to make
sure we're not doing that and only displaying them in priority order.

Original-Fixed: 124402
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845756
Original-Revision: 8d259d0fa0f19506c18984c323a8f6fdfa14ec0f
GitOrigin-RevId: b66bc4233a746cd115b1e26d772e3daecc82c3ac
Change-Id: I2452ec319f7089f47cef3a054bd436f7c1492998