[roll] Roll fuchsia [netstack3] add fuchsia.net.routes/State.Resolve
Implements the basic functionality needed for `Resolve`, including
looking up the destination in the routing table, identifying the
preferred source address for communicating with the destination, and
querying the neighbor table for the remote's MAC address.

A follow up CL will be needed to handle the case where the neighbor is
not present (or incomplete) in the neighbor table. Core will need to
initiate neighbor discovery, and bindings will need to wait for it to
finish.

Original-Bug: 120878
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822787
Original-Revision: f8b373a308b6964f5403acd6fdc4caf84cc0addc
GitOrigin-RevId: e3b87fe4b61d5d9ea32639ae85d920d0d9730235
Change-Id: Ie9817e639d79ecd0831c322626eecf020f8eb1d9