[roll] Roll fuchsia [sys] Rename CapabilitySourceInterface to CapabilitySource
CapabilitySource is an enum. While one could argue that
the word "Interface" makes sense for "ComponentInstanceInterface" since
that's a trait, that still doesn't follow the Rust naming convention.
The analogy is even less tenable for the CapabilitySource enum.

Is there a reasoning behind the "Interface" suffix? If not, it seems
less confusing to remove it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845741
Original-Revision: dc4cb515b326d18444b0af35781cdb831747b9ed
GitOrigin-RevId: 9d82cc09ff97e44a6f9bbb525a7e222bc1caebe3
Change-Id: I8a04ec977a7e3cf0b7c543fa89f0bb5fbaaa90e9