[roll] Roll fuchsia [display][virtio] Move driver to //src/graphics/display/drivers
The //src/graphics/drivers/virtio driver is the guest-mode
driver of virtio-gpu device; it is only used for framebuffer
scanout and not for any other graphics virtualization, so
this change moves the driver to //src/graphics/display/drivers.

To better distinguish the driver from the host-mode driver
of virtio-gpu in fuchsia (located in //src/virtualization),
we rename the driver directory to "virtio-guest" to reflect
this.

Test: no-op. fx set core.x64 ; fx build ; fx qemu
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845576
Original-Revision: e5d803c900be3b75629606ba9e696371be8abdfb
GitOrigin-RevId: d745c244e501491200566ea6f005582cf6d8196e
Change-Id: Ic3e9680e16b24417e73bb70cb85465a16e69c05a