[roll] Roll fuchsia [sys] Add comments to RoutingStrategy
And also add `<(), (), ()>` to the impl for the initial type state,
because the older syntax makes it look like the `new`, `use` methods
etc are composed into all the subsequent RoutingStrategy type states,
when in fact they're only available on the intial type state. We might
as well rename `RoutingStrategy::new` to a `routing_strategy()` free
function and get rid of the generic args defaulting -- that's another
option.

I'll let the reviewers decide if this is idiomatic Rust/clearer.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/846116
Original-Revision: 74a6203894a14bccfab41641b715ea40ffb9d8b8
GitOrigin-RevId: d2de9ffc07ab6ae87627f88f3b48a0a3a0788d7c
Change-Id: I03b1516a2f987fd2485c6af924ddba734ef2d85b