[roll] Roll fuchsia [kernel] Extend core test timeouts on RISC-V
The Bti.DecommitRace test case takes 15-17 minutes to run on RISC-V. To
get the bot green, this CL extends the timeout of the tests to 30
minutes. We chose 30 minutes as it is the default timeout (10 minutes)
plus the 17 minutes needed for the DecommitRace test case, along with a
few minutes of buffer.

Original-Bug: 126120
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/846853
Original-Revision: 1bdcdf3d6132625c5a67ede22f1139ef6b9cab69
GitOrigin-RevId: 416242ecb2a785590d5c860f4f81445bfc8950c2
Change-Id: I3865ceebf2e582f9d38c4de3058bb3741ec624a1