[roll] Roll infra/recipes [cmake] Add build_with_ninja() function
api.cmake.build_with_ninja() is a high-level utility for building with
CMake and Ninja. ~All our current uses of CMake are paired with Ninja
and use almost the exact same flow, aside from cmake configurations
through `extra_args` i.e.

1. Configure with CMake
2. Build with Ninja
3. ninja install

This is a proof-of-concept which adds the function and replaces the
logic in the zlib module (to start with). Anything that falls under the
above pattern should migrate to this module in the near future.

No functional change.

Original-Bug: 122100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/846776
Original-Revision: 44122fc3727def5f0c543b3d1020c45d565376e7
GitOrigin-RevId: b31516c27bd13c3a67a025655b3c4f2bf48b0630
Change-Id: Id2ae6b243163f74d3c8c88ea54aa633f5c5d84a9