[roll] Roll fuchsia [devicetree] Suppress -Wdangling-pointers warning
GCC 13's reports -Wdangling-pointers warnings for this code,
which uses a pattern that rightfully looks a bit suspect, but is
actually working as intended.  So just suppress the warning for
that code, along with using RAII semantics to make it more robust
in case any more `return` statements get added to the function.

Original-Bug: 126220
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847476
Original-Revision: 6127536dbcb2028045035e0fc348cda702af2154
GitOrigin-RevId: 48098fa80a31b9a5575cdbf3ee63707490c791d0
Change-Id: Ie5cbcec3464aecc892bafb668e4e48d6b8180b0e