[roll] Roll fuchsia [superproject] Roll scudo [scudo] Fix missing pushing 1 block to BatchClassId
This was happened rarely. The only case is when a thread is teared down
and it only has one block of BatchClass and the freelist of BatchClass
is empty. The impact is leaking 1 block of BatchClass and which is minor.

Differential Revision: https://reviews.llvm.org/D149141

GitOrigin-RevId: b4087b071b174c535be45bd13069c3b610f24dba
Original-Revision: 532769adf209c8bbc432c834f194295e94405785
Roller-URL: https://ci.chromium.org/b/8782672810706086961
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847008
Original-Revision: 7a4463709e0251a9f974d3948170fb6c8791a1e0
Change-Id: Ie72abe7318366bda5ae8807da7148e7f9fd8b61a