[roll] Roll fuchsia [starnix] Skip InterpreterScriptTests
These were expected to fail, but are moved to skip to migrate the tests
to the debian image.

Once the migration is complete, more tests will pass, but some will
still need to be moved from fail -> skip because of the failure mode.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847816
Original-Revision: 9bad731028fe1a1e0e06599fe5838afe84adfd0b
GitOrigin-RevId: 5c01a442a42f77aa62c007666b818342c16bc02f
Change-Id: I382390d0ee910113ac636fd49fa8f41b7f5d5f42