[roll] Roll fuchsia [kernel][diags] Clean up "dump_thread" names
Move most of the stand-alone dump_thread diagnostic routines so that
they are members of Thread, Thread::Current, or a newly introduced
ThreadDumper.

This rearrangement removes a few redundant declarations, and makes it
a bit easier to express the static locking requirements as we move to
break up the global thread lock.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847296
Original-Revision: 2e6ced08beafe71295fc259f5b837dfdabf7f8db
GitOrigin-RevId: 666cd15c2e8c4e791af56f2e2869dc9eed50a3d7
Change-Id: I6cb1bc7fc567e9697e765f8362a10db6701b07f3