[roll] Roll fuchsia [boot_test] Ensure disabled boot tests emit test metadata
Previously, we would not emit test metadata for disabled boot tests, as
the indicator for whether it was a boot test was lost (being specified
within test environment metadata); otherwise test machinery would assume
it was a normal test and would try to run the disabled test in the wrong
way. With recent metadata refactors this is no longer an issue, so we no
longer need to suppress test metadata. The main upshot of this is that
`fx run-boot-test` can now still identify and run disabled boot tests.

Original-Fixed: 122881
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/848096
Original-Revision: 9fe721e99a6203a5bdd0213ee1c628ced2874cb4
GitOrigin-RevId: 3ffa9e55c65da88bb8fb4548cd95064bd769a853
Change-Id: I81865117ed255ae7835d2e1948693bf7cc5b6489