[roll] Roll fuchsia [kernel][vm] Remove gen count optimization
The purpose of the gen count is to determine when attribution is
changed and is ambiguous then whether code paths that do not imapct
attribution should increment the gen count. Failure of those paths
to increment would cause this optimiziation to be bugger.

Due to the way the new LookupCursor works, re-performing the page
lookup is actually quite cheap now and this optimization is not
providing any value.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847276
Original-Revision: c214c77ad484bbd36beb76f2fa982562de355ed6
GitOrigin-RevId: 51fcf1ba9f46bda0d7e02cde14dff4ccedc371cf
Change-Id: Ib012fec2ae7933e4b72229fe870a163ba43463a0