[roll] Roll fuchsia [06/18][display][coordinator] Use an encoding allowing AnyPixelFormat for pixel_format and cursor_info.
Previously we stored the banjo encoding of pixel formats
(zx_pixel_format_t) and cursor info (cursor_info_t), and do static
assertion to make sure that the banjo formats are the same as their
FIDL equivalents (uint32_t and fuchsia.hardware.display.CursorInfo).

However, this is going to be not true when we migrate pixel formats
to not use zx_pixel_format_t.

This CL changes the types of pixel_format and cursor_info storage
throughout the display coordinator; now it stores an internal encoding
format (actually same as FIDL/banjo, with type aliases indicating that
the pixel formats are of "AnyPixelFormat" instead of an unclear
uint32_t). This CL adds banjo-to-coordinator conversion logic and
coordinator-to-fidl conversion to migration-util.h.

Currently they copy the values verbatimly, and in the future they'll
perform format conversions as specified by the FIDL / banjo protocol.

Test: no-op. display-coordinator-unittests
Original-Bug: 126114
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843242
Original-Revision: be3f673aff65b0ddb5db5cf40ba048c7802b269c
GitOrigin-RevId: f30fc2c2596e60787ff8850f21dbecffba21bc50
Change-Id: I3a2e7c1e587a7b2c362926bba3cacb219156da35