[roll] Roll fuchsia [build] Always use the local pool for non-remoted actions
Always use the 'local' pool for locally-executed actions, instead of
the full (10*num_cores) concurrency that is available when goma is
enabled.

Correct the logic that sets the use of the local pool for locally-
running rust actions when goma is enabled, but RBE is not.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847016
Original-Revision: 7a34ef93bad123f9c344bfeffd408762028d0093
GitOrigin-RevId: 66b10c17fa2baa4c408d6af06ef34321f44ee329
Change-Id: Id7289f5462f32b1fac3f236899f499bad55682ee