[roll] Roll fuchsia [fidl][rust] Replace table consts with fn or let
This CL replaces each const that contains a FIDL table with a function
returning the table or with a let binding, depending on the situation.
This is necessary for the migration from EMPTY to default(), because
default() is not a const fn. Such consts were fragile anyway, e.g. they
did not work with String fields for example.

Original-Bug: 102876
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845818
Original-Revision: 2ca0e742cc7c05bdb6d90ec19706769d8a483cf0
GitOrigin-RevId: e8e178265f13ac99826801a4e8bb6ca09ecb4756
Change-Id: Ibeb84d6b64473e282bc2e8572f51293384ec7637