[roll] Roll fuchsia [fuchsia-repo] Fix flakes by tracking removed manifests
Previously we didn't observe manifests that we were no longer interested
in watching, so we needed to poll for a period of time to see if no
events were received. This changes the PackageManifestWatcher to now
include the manifests we stopped watching, so the tests can check we
correctly stopped watching these manifests.

Original-Bug: 124412
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/848066
Original-Revision: e41563705807b60ddabb11c51981bf19ec91e1cc
GitOrigin-RevId: 81f7975b463984e9e63231f1da546bc951abb4f0
Change-Id: I293882a771cc52a979fbc0ac84c8e205aea87e0f