[roll] Roll fuchsia [check-licenses] Updates to readme package.
Previously, the tool would recursively search for license files
in go/dart/rust packages to add to the project. This strategy
disregards any skiplists or other configuration settings in other
parts of the tool, which is not good.

This CL changes it to only add any license files at the top-level
of those project types. If other licenses are found deeper in
a given project, they'll be added using a separate process.

Updated the tests to reflect this new behavior.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/848157
Original-Revision: ff65b41ed1b6d364ce427c807a5ed02eff47565e
GitOrigin-RevId: c3894f1dc43b612af0b028ddcad97930f6957366
Change-Id: I6321142f31ec1f97e97d026ff5a0941fd538f66d