[roll] Roll fuchsia [fidl][rust] Migrate away from new_empty()
This CL migrates Rust code away from using Decodable::new_empty(). This
function is used internally to create a valid instance to decode into.
It was never meant to be part of the public API.

Although it can be useful to make an "empty" or "zero" value, exposing
new_empty() is undesirable because that interpretation only makes sense
some of the time. For example, with a union or enum, new_empty()
arbitrarily picks the first variant.

I've made a few kinds of changes:

* For tables, I changed new_empty() to Default::default(). The Default
  trait was added in I0c1e6ff00ad438bc65e9b6e7872ab4cbfc7563dc.

* For other types, I expanded out the value that new_empty() returns.
  When it was used often, I factored it out into a const or function.

* In some cases new_empty() was used just to assign the fields
  imperatively. I changed these to use struct initializers.

* I removed all the fidl::encoding::Decodable imports.

* I changed some tests in src/connectivity/wlan to use event streams to
  check for an epitaph rather than manually decoding it off the channel.

Original-Bug: 118783
Original-Bug: 102876
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849236
Original-Revision: e1b8506b2e729278430049e223b76ff560269e8a
GitOrigin-RevId: c9a69e8d38a8c2ce853e255528a3e9e9b05d11af
Change-Id: I5d15dac6db4a63ce0fdc4ed67f1984732b0dea2f