[roll] Roll fuchsia [netstack] Fix memory leak when CreateNIC fails
Fix a memory leak where if `stack.CreateNICWithOptions` fails, the
qdisc that was initialized and passed as argument was not getting
cleaned up, and the goroutine(s) spawned and the memory allocated
for the queue(s) would leak. This has been fixed by calling `Close`
on the qdisc, which stops the goroutine(s) and allows the queues to
be garbage collected.

Original-Fixed: 126332
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849196
Original-Revision: 6194cabce369ccffcca88a370c033fa13f687690
GitOrigin-RevId: 2d84d4a8e3b47c531e36d8f201904825da591d17
Change-Id: Ied4c6a4196f21e250307ff516e05ab59999a4fd0