[roll] Roll fuchsia [virtio-guest] Rework VirtIO ABI to match C++ style guide.
The current ABI definition used by the virtio display driver is a strict
subset of the VirtIO definitions, which are provided in C.

This CL reimplements the ABI definition to match the readability
standards that we've set for our register definitions. This is the
relevant bar, because the ABI is our interface with the virtualized
hardware, similarly to how the registers are our interface with
(usually) physical hardware.

Test: fx test virtio-gpu-test
Test: fx qemu --virtio -g -N -q /usr/bin
Test: fx shell display-tool squares
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/848358
Original-Revision: 66292531144ff44c89a630b7acac8f45888ba77c
GitOrigin-RevId: b6ac4554f1daa7c486191f5ab6693578de0d111c
Change-Id: Ie49c603cc3c84a4a26739aefdeb3029fd2472572