[roll] Roll fuchsia [03/14][fidl][display] Remove ImageConfig.pixel_format.
This field is not used in any in-tree or out-of-tree code.

Since it modifies the struct, this is an API and ABI-breaking change.

Currently the only out-of-tree usage of the display API is through
VkLayer_image_pipe_swapchain_fb; it's not automatically tested so
we'll need to manually rollout the layer binary.

Test: fx build
Original-Bug: 126113
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843241
Original-Revision: 569539a7a1a4f52706242fb4872948880806f345
GitOrigin-RevId: b0deab1bb1bb3592a776381d36a0d2105b319be9
Change-Id: I764079ffb24dac98f06a41834b9af3d7a3c738e9