[roll] Roll fuchsia [07/14][banjo][display] Remove pixel_format from Image.
Now that image_t.pixel_format is not used by drivers nor
display coordinator anymore, it's safe to delete the field.

This is only used in-tree and it's not exposed as part of SDK,
so this change is both API and ABI safe.

Test: fx build
Original-Bug: 126113
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843239
Original-Revision: a4d52d668805ada0a0bb97deb2da1951f2fdbecb
GitOrigin-RevId: 9609086f22e7894cc576a0083e3b8afc6f2b9609
Change-Id: I81eaa7e1935e3bacfc5dc798e094ea7d2d3433d4