[roll] Roll llvm-project/libc [libc] Add support for global ctors / dtors for AMDGPU
This patch makes the necessary changes to support calling global
constructors and destructors on the GPU. The patch in D149340 allows the
`lld` linker to create the symbols pointing us to these globals. These
should be executed by a single thread, which is more difficult on the
GPU because all threads are active. I chose to use an atomic counter to
sync every thread on the GPU. This is very slow if you use more than a
few thousand threads, but for testing purposes it should be sufficient.

Depends on D149340 D149363

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D149398

GitOrigin-RevId: e6907e1dff05e76f1b101226bfbf24484e4df73e
Original-Revision: ec040b7f0382eb47cda869f69d4bca02fc872b25
Change-Id: Iccb8377ccd259c3a23f18a42014639258d735070