[roll] Roll fuchsia [wlan][tool] fix saving network with password only
This fixes the case where a user specifies an SSID and password to save
a network but doesn't specify the security type or credential type. WPA2
is used by default because most users will have WPA2 or WPA3 networks,
and a network saved with WPA2 can be used to connect to WPA2 or WPA3.

Test: added unit test (fx test donut-lib-tests)
Original-Bug: 126081
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849396
Original-Revision: 322b084362b966bfd19d82b43673011cc067599b
GitOrigin-RevId: 87a7a74cac6d7a9afe32a6520b01998bf5edab9e
Change-Id: Ib9e377a43a5f798ba0da520b3a1b6e7c34611801