[roll] Roll fuchsia [netstack] Introduce Diagnostics.GetProcessHandle
Add a method to the fuchsia.net.debug/Diagnostics protocol to allow
callers who have the capability to get a handle to the netstack's
process for debugging and diagnostic purposes. Note that the process
handle is only vended with the ZX_RIGHT_INSPECT right to limit the
amount of privilege given to the caller.

Original-Bug: 118917
Test: netstack-debug-fidl-test --test-filter get_process_handle_*
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849296
Original-Revision: e1236adc5958574f9517e089493d4074b7e8d8ad
GitOrigin-RevId: 9677034a70dd1f9dc2d280ba80e428eecc6cadba
Change-Id: I3583037094acc646cfa5b25f647ef5dbe855be67