[roll] Roll fuchsia [banjo] Notify the driver runtime of banjo call
This change introduces driver context changes from banjo calls
into the driver runtime environment. This is done by storing the owning
driver in the server side of protocols and interfaces, and then setting
and desetting it in the runtime. This is an optional feature that can be
enabled by passing true into the *Protocol/*Interface template. This
is also working for async callbacks. ddk-callbacks are not supported.

Original-Bug: 116791
Tests: Manually tested on a sherlock by booting it and running all the
tests under //src/devices:tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797946
Original-Revision: 886608fcb6b44f60b529b28b6f42c34be85b6dd8
GitOrigin-RevId: 76444a5418440edd13cce91bf8cff7043f96a4e9
Change-Id: I8ea21d6c85bfabb8d46382a03398bdfd8baae66d