[roll] Roll fuchsia [ffx][emu]Making the instance name default configurable
This makes the default emulator instance name configurable by setting
`emu.name`. The intent of this feature is to allow build systems or
similar, to set the configuration based on the product/board being run
in order to make it more clear what "type" of emulator is running, and
make it easier to serve the correct packages to the target instance.

Original-Bug: 125893
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849600
Original-Revision: f6b47298aac8f549fba38f2b0e5fd03e90c647ba
GitOrigin-RevId: accbe348624a654c158f71b66bbd27462f6d5942
Change-Id: I48c910dfe881d9c6a7aa2800c5328ee1fb89eb10