[roll] Roll fuchsia [flatland] Ignore rows that may be invalid on sherlock / vim3.
On sherlock and vim3 sometimes the last few rows (on sherlock it's the
last 5 rows, on VIM3 with 1920-width display it's the last 2 rows) can
be missing from the captured bytes.

This change modifies the comparison code so that it ignores rows that
can be invalid.

Original-Bug: 125394
Original-Bug: 125842
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850120
Original-Revision: 96edeb19cd663d1af7f49d559e7689691362d4b9
GitOrigin-RevId: a617528d14adf680f426f761bd8565e1b99f2d9a
Change-Id: I08f281586c5a22b957db1a8b9cc7a1ca392bc489