[roll] Roll fuchsia [virtualizaton] Build test_util binaries in CQ
These are prone to build regressions as a result of FIDL refactorings.
Without this CL, the virtualization builders would need to be added to a
CL to discover that it regressed the builds of these binaries.

This has been a frequent cause of failures in the virtualization
builders and this CL seeks to surface these errors in CQ without adding
substantial load on infra bots or build image size.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850479
Original-Revision: 7682226d018d7cbacefc07833a1d992ae1e29d79
GitOrigin-RevId: 4b8d0da320fc3b4336ab66d6e16f3388f3219c4b
Change-Id: I3c3f42389072d96bf8a6862a60bb14b3a1a702e6