[roll] Roll fuchsia [fxfs] Add seek table to persistent layer.
Add a seek table to the end of persistent layers that gives the id of
the first part of the search key, allowing for fast approximate search
of the entire layer.

In order to support this id searching method, a trait is added to the
Key trait which documents and requires that each key begin with a u64
that will be the first deciding faster in sorting the overall Key.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838058
Original-Revision: 688fa2053a5013a43a5b076bb48f24b86f64841c
GitOrigin-RevId: 807f0dcce2d7e93dba40b0dc8bd23074b528a969
Change-Id: I5b7bbc0577643716897dca4582ecdbd8f26dc10e