[roll] Roll fuchsia [product] conditionally build product bundle v1
In preparation for removing PBv1, add a flag to avoid building PBv1.
Right now, this defaults to True, so this CL is somewhat of a no-op.
In the future, this will default to False when we believe there
shouldn't be any issue in doing so, but we can still quickly re-enable
it.
In the further future, the PBv1 generation will be removed and these
markers (flag/config checks) will help guide what to delete.

Note: this CL doesn't necessarily remove all the PBv1 functionality.
I.e. I expect to add more checks as issues are found.

Original-Bug: 126216
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847436
Original-Revision: 23e59efce790f52714bd3e37f7e0bb7f25ff1203
GitOrigin-RevId: 7254b41ddde56690904abd7378452d3a829374f1
Change-Id: I89e2f03716e08c0c55eb7448ee7cb40b1ea1bac0