[roll] Roll fuchsia [wlan][mlme] Add missing binding for scan request arguments
The DeviceInterface requires passing a pointer to a Banjo data
structure, but the memory backing that pointer should not go
out of scope. This change binds the memory to a local variable that
remains in scope through the entire function call.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851220
Original-Revision: abe31b09d101224fb12532ab3f5bcdd7a315b565
GitOrigin-RevId: 1e4f467c3497c8d26f082538d6c79620500ed30e
Change-Id: I8ebaae6c86d0339045e93c0aabc98b0ead548426