[roll] Roll fuchsia [ffx][doctor] Check SSH Key validity
This adds a check to the doctor that checks that the SSH keys
exist, and that the private key and authorized keys are consistent.

It also adds a command line switch to repair or regenerate the keys if
needed.

Original-Bug: 113497
Original-Bug: 124582
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841297
Original-Revision: 1baa821b8c78d4c91f54cccdce0271b99c184265
GitOrigin-RevId: 9f434ddefae94806d2a0898254484f472f91e25a
Change-Id: I5707e3f3b1deea56d813a6168482d345c7162427