[roll] Roll fuchsia [rbe] Write download stubs, download later
When --download_outputs=false, write out download stubs.
Stubs are set up as .dl-stub files, symlinked from the
file's actual path.

A stubbed download for 'foo/bar.o' will look like:
  foo/ (dir)
    \- bar.o (symlink -> bar.o.dl-stub)
    \- bar.o.dl-stub (text file)

On download, the link to the stub file is followed,
the blob is downloaded to a temp file.
If the download succeeds, the link is replaced with the full file.
The stub file (.dl-stub) is kept around, to make it
easier to inspect the origin.

Original-Bug: 123178

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841919
Original-Revision: 3cbe243d99d58267088b22640b4b8df378fe9c1f
GitOrigin-RevId: 5b862f1cd2387284fd14f8b4c65734676301c33b
Change-Id: Iafebe210d3930b2630a6d9b48d95a1c73e7ce2ca