[roll] Roll fuchsia [cf] Add config types to fuchsia.component.decl.
These types are ABI-compatible with the types in
fuchsia.component.config so that migration between the
two won't cause any breaking changes.

Removes the fostr formatter for fuchsia.component.decl
because it is not compatible with the Value types and
would be too much effort to fix when that library is
only used in one test.

Original-Bug: 96254
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850816
Size-Review: James Sullivan <jfsulliv@google.com>
Original-Revision: 636dea9d99deea0e3723c78e3c9694ee5e090562
GitOrigin-RevId: e1a9a1f45d74c270c0fa7c8e520d99881106e6f2
Change-Id: I9e90be5595570068b036445e10e576b9de08ee4c