[roll] Roll fuchsia [rbe] download tool command prefix
For commands that might need to use remote outputs that
were not downloaded due to rewrapper --download_outputs=false:

  build/rbe/dlwrap.py --download ... -- command

or

  build/rbe/dlwrap.py --download ... && command

(depending on which is more suitable).

--download arguments are paths that point to places that would
normally would be real files, but could be download stubs.
Real files are ignored and left as-is, stubs (links) are downloaded
and written over the same locations.

Note: this does not require any xattrs because reproxy is not involved.

Original-Bug: 123178
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851319
Original-Revision: b5a503d355422b4a1a4baf33bfac18fbe520aa0d
GitOrigin-RevId: 761987b0f1e3a9be0efba822b0b0e3703e87fde4
Change-Id: Idd02f4fe25eed133dcc722281e6a470b9534d0e4