[roll] Roll fuchsia [testrunner] Fix nsjail env variable handling
The nsjail wrapper is splitting on `=` and is mutilating environment
variables that include a `=` in the value. Do a strings#Cut instead to
only split on the first `=`, ignore `found` as os.Environ() is
guaranteed to return strings containing the separator.

Original-Bug: 88169
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851416
Original-Revision: 56b8e08f841dc3486836288af2db7af438968a1b
GitOrigin-RevId: ebeb461ebaa23818363ba444e09daae551a7dce3
Change-Id: I68436a5b4d27a5f0738a298a0cdfaaaced1d69a7