[roll] Roll fuchsia [dfv1] Use MatchDriver instead of MatchDriversV1
Use the DFv2 version of matching which will return an error if we have
multiple matches. This should be possible now that DriverManager is
requesting specific bind urls.

Print better error messages in driver index while we're here.

This removes the V1 API. A number of duplicate DriverIndex tests
were removed, others were updated to use the new API.

driver_loader_test was also updated to use the new API.
A few of the tests are no longer relevant, because DriverIndex now
has full control over things like fallback drivers.

The ddk-fallback-test for eager binding was removed. This is because if
you set a fallback driver as non-fallback, it will conflict with the
other non-fallback driver and nothing will bind. Not binding anything
is preferred, because in V1 whichever driver "happened" to be first
would win, and the order was alphabetical.

Original-Fixed: 81175
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847036
Original-Revision: 9d2a162042098874315a25377c3fbb4abcf3b2cb
GitOrigin-RevId: 8aa4e32a6bba25eb2afd5c305fddb4896d64a473
Change-Id: Idd4754152b457708f310cd8cc81c06584ecdefbc